refactor column comparison predicate (#3373)
