+++ be/src/storage/vectorized/column_predicate.h
@@ -52,6 +52,8 @@ enum class PredicateType {
     kMap = 15,
 };
 
+std::ostream& operator<<(std::ostream& os, PredicateType p);
+
 inline TExprOpcode::type convert_predicate_type_to_thrift(PredicateType p) {
     switch (p) {
     case PredicateType::kEQ:
@@ -211,116 +213,14 @@ protected:
 
 using PredicateList = std::vector<const ColumnPredicate*>;
 
-template <template <FieldType> typename Predicate, template <FieldType> typename BinaryPredicate>
-ColumnPredicate* new_column_predicate(const TypeInfoPtr& type_info, ColumnId id, const Slice& operand) {
-    auto type = type_info->type();
-    switch (type) {
-    case OLAP_FIELD_TYPE_BOOL:
-        return new Predicate<OLAP_FIELD_TYPE_BOOL>(type_info, id, string_to_int<int>(operand) != 0);
-    case OLAP_FIELD_TYPE_TINYINT:
-        return new Predicate<OLAP_FIELD_TYPE_TINYINT>(type_info, id, string_to_int<int8_t>(operand));
-    case OLAP_FIELD_TYPE_UNSIGNED_TINYINT:
-        return nullptr;
-    case OLAP_FIELD_TYPE_SMALLINT:
-        return new Predicate<OLAP_FIELD_TYPE_SMALLINT>(type_info, id, string_to_int<int16_t>(operand));
-    case OLAP_FIELD_TYPE_UNSIGNED_SMALLINT:
-        return nullptr;
-    case OLAP_FIELD_TYPE_INT:
-        return new Predicate<OLAP_FIELD_TYPE_INT>(type_info, id, string_to_int<int32_t>(operand));
-    case OLAP_FIELD_TYPE_UNSIGNED_INT:
-        return nullptr;
-    case OLAP_FIELD_TYPE_BIGINT:
-        return new Predicate<OLAP_FIELD_TYPE_BIGINT>(type_info, id, string_to_int<int64_t>(operand));
-    case OLAP_FIELD_TYPE_UNSIGNED_BIGINT:
-        return nullptr;
-    case OLAP_FIELD_TYPE_LARGEINT:
-        return new Predicate<OLAP_FIELD_TYPE_LARGEINT>(type_info, id, string_to_int<int128_t>(operand));
-    case OLAP_FIELD_TYPE_FLOAT:
-        return new Predicate<OLAP_FIELD_TYPE_FLOAT>(type_info, id, string_to_float<float>(operand));
-    case OLAP_FIELD_TYPE_DOUBLE:
-        return new Predicate<OLAP_FIELD_TYPE_DOUBLE>(type_info, id, string_to_float<double>(operand));
-    case OLAP_FIELD_TYPE_DISCRETE_DOUBLE:
-        return nullptr;
-    case OLAP_FIELD_TYPE_DECIMAL: {
-        decimal12_t value;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DECIMAL>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_DECIMAL_V2: {
-        DecimalV2Value value;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DECIMAL_V2>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_DECIMAL32: {
-        int32_t value;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DECIMAL32>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_DECIMAL64: {
-        int64_t value;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DECIMAL64>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_DECIMAL128: {
-        int128_t value;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DECIMAL128>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_DATE: {
-        uint24_t value = 0;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DATE>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_DATE_V2: {
-        int32_t value = 0;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DATE_V2>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_DATETIME: {
-        uint64_t value = 0;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_DATETIME>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_TIMESTAMP: {
-        int64_t value = 0;
-        auto st = type_info->from_string(&value, operand.to_string());
-        DCHECK(st.ok());
-        return new Predicate<OLAP_FIELD_TYPE_TIMESTAMP>(type_info, id, value);
-    }
-    case OLAP_FIELD_TYPE_CHAR:
-        return new BinaryPredicate<OLAP_FIELD_TYPE_CHAR>(type_info, id, operand);
-    case OLAP_FIELD_TYPE_VARCHAR:
-        return new BinaryPredicate<OLAP_FIELD_TYPE_VARCHAR>(type_info, id, operand);
-    case OLAP_FIELD_TYPE_STRUCT:
-    case OLAP_FIELD_TYPE_ARRAY:
-    case OLAP_FIELD_TYPE_MAP:
-    case OLAP_FIELD_TYPE_UNKNOWN:
-    case OLAP_FIELD_TYPE_NONE:
-    case OLAP_FIELD_TYPE_HLL:
-    case OLAP_FIELD_TYPE_OBJECT:
-    case OLAP_FIELD_TYPE_PERCENTILE:
-    case OLAP_FIELD_TYPE_JSON:
-    case OLAP_FIELD_TYPE_MAX_VALUE:
-        return nullptr;
-        // No default to ensure newly added enumerator will be handled.
-    }
-    return nullptr;
-}
-
 ColumnPredicate* new_column_eq_predicate(const TypeInfoPtr& type, ColumnId id, const Slice& operand);
 ColumnPredicate* new_column_ne_predicate(const TypeInfoPtr& type, ColumnId id, const Slice& operand);
 ColumnPredicate* new_column_lt_predicate(const TypeInfoPtr& type, ColumnId id, const Slice& operand);
 ColumnPredicate* new_column_le_predicate(const TypeInfoPtr& type, ColumnId id, const Slice& operand);
 ColumnPredicate* new_column_gt_predicate(const TypeInfoPtr& type, ColumnId id, const Slice& operand);
 ColumnPredicate* new_column_ge_predicate(const TypeInfoPtr& type, ColumnId id, const Slice& operand);
+ColumnPredicate* new_column_cmp_predicate(PredicateType predicate, const TypeInfoPtr& type, ColumnId id,
+                                          const Slice& operand);
 
 ColumnPredicate* new_column_in_predicate(const TypeInfoPtr& type, ColumnId id,
                                          const std::vector<std::string>& operands);
@@ -330,23 +230,5 @@ ColumnPredicate* new_column_null_predicate(const TypeInfoPtr& type, ColumnId, bo
 
 ColumnPredicate* new_column_dict_conjuct_predicate(const TypeInfoPtr& type_info, ColumnId id,
                                                    std::vector<uint8_t> dict_mapping);
-
-template <FieldType field_type, template <FieldType> typename Predicate, typename NewColumnPredicateFunc>
-Status predicate_convert_to(Predicate<field_type> const& input_predicate,
-                            typename CppTypeTraits<field_type>::CppType const& value,
-                            NewColumnPredicateFunc new_column_predicate_func, const ColumnPredicate** output,
-                            const TypeInfoPtr& target_type_info, ObjectPool* obj_pool) {
-    using ValueType = typename CppTypeTraits<field_type>::CppType;
-    const auto to_type = target_type_info->type();
-    if (to_type == field_type) {
-        *output = &input_predicate;
-        return Status::OK();
-    }
-    auto cid = input_predicate.column_id();
-    const auto type_info = input_predicate.type_info();
-    std::string str = type_info->to_string(&value);
-    *output = obj_pool->add(new_column_predicate_func(target_type_info, cid, str));
-    return Status::OK();
-}
 
 } //namespace starrocks::vectorized
