+++ be/src/storage/vectorized/zone_map_detail.h
@@ -1,5 +1,9 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
 
+#pragma once
+
+#include "column/datum.h"
+
 namespace starrocks {
 namespace vectorized {
